Added some stuff for armies
