+++ Sprint1/src/com/team7/objects/unit/combatUnit/CombatUnit.java
@@ -5,13 +5,5 @@ import com.team7.objects.unit.Unit;
 
 public abstract class CombatUnit extends Unit {
 
-    private Army army;
 
-    public Army getArmy() {
-        return army;
-    }
-
-    public void setArmy(Army army) {
-        this.army = army;
-    }
 }
