+++ Sprint1/src/com/team7/objects/unit/nonCombatUnit/Colonist.java
@@ -10,6 +10,7 @@ public class Colonist extends NonCombatUnit {
 	    UnitStats stats = new UnitStats(0, 0, 0, 5, 100, 1);
         int id = ProbabilityGenerator.randomInteger(0, 99999);
 
+		setArmy(null);
         setUnitStats(stats);
         setPowered(true);
 	    setId(id);
